+++ src/TestSuiteStorage/Action/RouteTestContract.php
@@ -13,6 +13,7 @@ use Heptacom\HeptaConnect\Storage\Base\Action\Route\Create\RouteCreatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Create\RouteCreatePayloads;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Create\RouteCreateResult;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Create\RouteCreateResults;
+use Heptacom\HeptaConnect\Storage\Base\Action\Route\Delete\RouteDeleteCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Find\RouteFindCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Find\RouteFindResult;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Get\RouteGetCriteria;
@@ -20,6 +21,7 @@ use Heptacom\HeptaConnect\Storage\Base\Action\Route\Get\RouteGetResult;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Listing\ReceptionRouteListCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Action\Route\Listing\ReceptionRouteListResult;
 use Heptacom\HeptaConnect\Storage\Base\Bridge\Contract\StorageFacadeInterface;
+use Heptacom\HeptaConnect\Storage\Base\Exception\NotFoundException;
 use Heptacom\HeptaConnect\TestSuite\Storage\Fixture\Dataset\EntityA;
 use Heptacom\HeptaConnect\TestSuite\Storage\Fixture\Dataset\EntityB;
 use Heptacom\HeptaConnect\TestSuite\Storage\Fixture\Dataset\EntityC;
@@ -38,6 +40,7 @@ abstract class RouteTestContract extends TestCase
         $receptionListAction = $facade->getReceptionRouteListAction();
         $findAction = $facade->getRouteFindAction();
         $getAction = $facade->getRouteGetAction();
+        $deleteAction = $facade->getRouteDeleteAction();
 
         $portalNodeCreateResult = $portalNodeCreate->create(new PortalNodeCreatePayloads([
             new PortalNodeCreatePayload(PortalA::class),
@@ -76,8 +79,9 @@ abstract class RouteTestContract extends TestCase
             $testCreateResults = new RouteCreateResults($testCreateResults->filter(
                 static fn (RouteCreateResult $r): bool => !$r->getRouteKey()->equals($findResult->getRouteKey())
             ));
+            $routeGetCriteria = new RouteGetCriteria(new RouteKeyCollection([$findResult->getRouteKey()]));
             /** @var RouteGetResult[] $getResults */
-            $getResults = \iterable_to_array($getAction->get(new RouteGetCriteria(new RouteKeyCollection([$findResult->getRouteKey()]))));
+            $getResults = \iterable_to_array($getAction->get($routeGetCriteria));
 
             static::assertCount(1, $getResults);
 
@@ -94,6 +98,16 @@ abstract class RouteTestContract extends TestCase
                     static fn (ReceptionRouteListResult $r): bool => $r->getRouteKey()->equals($findResult->getRouteKey())
                 );
                 static::assertCount(0, $receptionListResult);
+            }
+
+            $deleteAction->delete(new RouteDeleteCriteria($routeGetCriteria->getRouteKeys()));
+
+            static::assertEmpty(\iterable_to_array($getAction->get($routeGetCriteria)));
+
+            try {
+                $deleteAction->delete(new RouteDeleteCriteria($routeGetCriteria->getRouteKeys()));
+                static::fail('This should have been throwing a not found exception');
+            } catch (NotFoundException $exception) {
             }
         }